move cuppy an and change and add display none to invisible
