+++ frontend/src/components/Cuppy/CuppySpeak.vue
